+++ src/app/page.module.css
@@ -1,13 +1,8 @@
-/*
-* #6697B6 Light Sky Blue
-* #527992 Medium Sky Blue
-*/
-
 
 .page {
   display: flex;
   flex-direction: column;
-  padding: 2rem 2rem 0 2rem;
+  padding: 1rem 2rem 0 2rem;
   min-height: 100vh;
   background-color: white;
   color: #201E20;
@@ -17,25 +12,13 @@
 .header {
   display: flex;
   flex-direction: column;
-  gap: 1.5rem;
-  margin-bottom: 1.5rem;
+  gap: 1rem;
+  margin-bottom: 1rem;
 }
 
 .navbar {
   display: flex;
   justify-content: space-around;
-}
-
-@media screen and (min-width: 600px) {
-  .navbar {
-    justify-content: center;
-    gap:25rem;
-  }
-}
-
-.elderName {
-  font-weight: bold;
-  color: #000046;
 }
 
 .appTitle {
@@ -52,14 +35,46 @@
   font-weight: bold;
 }
 
-.entryContainer {
+/*Main Content*/
+.main {
   display: flex;
   flex-direction: column;
-  gap: 0.5rem;
+  height: 75vh;
+  align-items: center;
+}
+
+.phoneNumberContainer {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+}
+
+.inputfield {
+  display: block;
+  margin-top: .5rem;
+  padding: .5rem;
+  font-size: 1rem;
+  width: 100%;
+}
+
+.label {
+  font-weight: bold;
+  color: #000046;
+}
+
+.CTAMessage {
+  color: #1CB5E0;
+  flex: .2;
+}
+
+.emotionalMessage {
+  color: #000046;
+  font-style: italic;
+  flex: 1;
 }
 
 .entryButton {
-  width: 50%;
+  width: 75%;
   align-self: center;
   padding: 0.5rem;
   display: flex;
@@ -69,9 +84,8 @@
   color: white;
   border-radius: 15px;
   cursor: pointer;
-  transition: all 300ms cubic-bezier(.23, 1. 0.32,1);
   -webkit-user-select: none;
-  margin-bottom: 0.5rem;
+  margin-top: 1.5rem;
   touch-action: manipulation;
   background: #000046;  /* fallback for old browsers */
   background: -webkit-linear-gradient(to right, #1CB5E0, #000046);  /* Chrome 10-25, Safari 5.1-6 */
@@ -88,136 +102,4 @@
 .entryButton:active {
   box-shadow: none;
   transform: translateY(0);
-}
-
-.entryButtonImageStyle {
-  height: 1rem;
-  width: 1rem;
-}
-
-.entryButtonMessage {
-  font-size: 14px;
-  font-style: italic;
-  text-align: center;
-}
-
-@media screen and (min-width: 600px) {
-  .entryButton {
-    align-self: flex-start;
-    width: 20%;
-    max-width: 175px;
-    min-width: 150px;
-    margin: auto;
-  }
-
-  .entryContainer {
-    align-items: center;
-  }
-}
-
-/*Content Styles*/
-.main {
-  display: flex;
-  flex-direction: column;
-}
-
-.contentTitleContainer {
-  display: flex;
-  justify-content: space-evenly;
-}
-
-.contentTitle {
-  font-weight: bold;
-  text-decoration: underline;
-  margin-bottom: 1rem;
-  text-align: center;
-  text-decoration-color: #527992;
-  list-style: none;
-}
-
-.contentTitle:hover {
-  color: #1CB5E0;
-  cursor: pointer;
-}
-
-.content {
-  height: 75vh;
-  overflow-y: auto;
-  overflow: scroll;
-  scrollbar-color: black; /*Works on firefox*/
-}
-
-/*Works on Chrome, Edge, and Safari*/
-.content::-webkit-scrollbar {
-  width: 10px;
-}
-
-/*Works on Chrome, Edge, and Safari*/
-.content::-webkit-scrollbar-thumb {
-  background-color: #527992;
-}
-
-.interactionDate {
-  font-weight: bold;
-}
-
-.interactions > p {
-  margin: .5rem 0;
-}
-
-.interactions h1 {
-  border: 1px solid #527992;
-  width: 50%;
-  margin: auto;
-  opacity: .4;
-}
-
-.writerName {
-  font-style: italic;
-}
-
-.titleFormatting::before {
-  content: "\a";
-  white-space: pre;
-}
-
-@media screen and (min-width: 600px) {
-  .interactions h1 {
-    width: 80%;
-  }
-
-  .content {
-    min-width: 600px;
-    margin: 0 auto;
-  }
-
-  .contentTitleContainer {
-    justify-content: center;
-    gap: 15rem;
-  }
-}
-
-.taskHeader {
-  font-weight: bold;
-  color: #1CB5E0;
-}
-
-.tasks {
-  padding-left: 1rem;
-  margin-bottom: .5rem;
-}
-
-.tasks::before {
-  content:"❤";
-  margin-right: .5rem;
-  color: #B85042;
-}
-
-.essentials {
-  margin-bottom: .5rem;
-}
-
-.highlight {
-  color: #1CB5E0;
-  text-decoration-color: #1CB5E0;
 }
